TestSubmission
(The only issue I had was forcing the discounted price to recalculate for products with a discount, as the components appeared to retain the previous discount of the item in that position. I would try parsing this discounted value through as a 'prop' for the component, but had trouble achieving this.

Otherwise I believe I have completed the test to the best of my abilities!)

And apologies again for the late submission.

Many thanks and kind regards.
- Andrew Walters